+++ src/main/java/org/linhart/ppj/sem/controllers/StateController.java
@@ -2,6 +2,8 @@ package org.linhart.ppj.sem.controllers;
 
 
 import org.linhart.ppj.sem.entities.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -12,6 +14,7 @@ public class StateController {
     private final StateRepository stateRepository;
     private final CityRepository cityRepository;
     private final MeteorDataRepository meteorDataRepository;
+    final private Logger logger = LoggerFactory.getLogger(CityController.class);
 
     public StateController(StateRepository stateRepository, CityRepository cityRepository, MeteorDataRepository meteorDataRepository) {
         this.stateRepository = stateRepository;
@@ -32,18 +35,27 @@ public class StateController {
     @PutMapping( "/states")
     public String addState(@RequestParam String stateName,@RequestParam String stateIso){
         stateRepository.save(new State(stateIso,stateName));
-        return "Added state "+stateName+" ["+stateIso+"]";
+        logger.info("Pridan stat {} [{}]", stateName, stateIso);
+        return "Pridan stat "+stateName+" ["+stateIso+"]";
     }
     @DeleteMapping("/states")
     public String deleteState(@RequestParam String stateIso){
         State state = stateRepository.findStateByIsoCodeEquals(stateIso);
-        List<City> cities = cityRepository.findCitiesByStateIsoEquals(state);
-        for (City city : cities) {
-            List<MeteorData> data = meteorDataRepository.findMeteorDataByCityIDEquals(city);
-            meteorDataRepository.deleteAll(data);
-            cityRepository.delete(city);
+        if (state != null) {
+            List<City> cities = cityRepository.findCitiesByStateIsoEquals(state);
+            for (City city : cities) {
+                List<MeteorData> data = meteorDataRepository.findMeteorDataByCityIDEquals(city);
+                meteorDataRepository.deleteAll(data);
+                cityRepository.delete(city);
+                logger.info("Odstraneno mesto {} ze statu {}", city.getName(), state.getIsoCode());
+                return "Odstraneno mesto " + city.getName() + " ze statu " + state.getIsoCode();
+            }
+            stateRepository.delete(state);
+            logger.info("Odstranen stat {} [{}]", state.getStateName(), stateIso);
+            return "Odstranen stat "+state.getStateName()+" ["+stateIso+"]";
+        }else{
+            logger.error("Stat se nepodarilo odstranit");
+            return "Stat se nepodarilo odstranit";
         }
-        stateRepository.delete(state);
-        return "Deleted state "+state.getStateName()+" ["+stateIso+"]";
     }
 }
